refactor page layout/routing
