+++ src/pages/App.tsx
@@ -5,6 +5,8 @@ import { HomePage } from './Home';
 import { Provider } from 'react-redux';
 import store from '../ducks/store';
 import Header from '../containers/Header';
+import Message from './Message';
+import Login from './Login';
 
 const App: FC = () => {
     return (
@@ -14,6 +16,8 @@ const App: FC = () => {
                     <Header />
                     <Switch>
                         <Route exact path='/' component={HomePage} />
+                        <Route path='/message' component={Message} />
+                        <Route path='/login' component={Login} />
                     </Switch>
                 </Provider>
             </BrowserRouter>
